Wait and retry for ubi device and volume nodes
There is a possibility of attempting mount operation
on a ubi-volume node even before it is created. To make
sure this doesn't happen, send an additional argument to
FindAndAttachUBI(), which represents the number of ubi
volumes present on a ubi device.
After performing a ubiattach on mtd device:
  1) Explicitly check for ubi-device node i.e.
     /dev/ubiX to be created.
  2) Then check for each ubi volume node i.e.
     /dev/ubiX_Y to be created.
If any of the above nodes are not created within a timeout
period, reboot the device to exit recovery mode.
This will make sure that we don't do partial upgrade (of one
image but not another because ubi devices were not created).
Once booted into normal mode, one can always trigger OTA upgrade again.
Additionally, redirect all logging to /dev/kmsg explicitly.

Change-Id: I2cb4d4bb5331c4588d11b0f92dca0ef3113cf674