Enable generation of ota packages for nand/ubi targets
This is a combination of 2 commits:

1) gen_buildprop is a task that modifies rootfs.
But it is executed as part IMAGE_PREPROCESS_COMMAND instead
of ROOTFS_POSTPROCESS_COMMAND. So, those tasks
that have a dependency on do_rootfs() will not see the
correct rootfs content. For instance, do_recovery_ubi()
misses out all files that are created as part of
gen_buildprop().
Execute gen_buildprop() as part of ROOTFS_POSTPROCESS_COMMAND.

2) Move do_gen_otazip_ubi() out of python
as build time variables are being not set inside
the python task.

Change-Id: Ibaabaeb10226c101776f857efea87b838bab51b0