Decouple recoveryfs generation from A/B OTA package generation
Currently, enabling RECOVERY_IMAGE for a target would trigger
generation of recoveryfs image and Non A/B OTA package as well.
However, on targets that support A/B boot, recoveryfs image is
not needed; A/B OTA package still needs to be generated though.

Introduce a new flag 'GENERATE_AB_OTA_PACKAGE' which when enabled
will not generate recoveryfs image but just an A/B OTA package.
The flag RECOVERY_IMAGE and GENERATE_AB_OTA_PACKAGE are set based on
ab-boot-support DISTRO_FEATURE. If ab-boot-support is present,
RECOVERY_IMAGE is set to '0' and GENERATE_AB_OTA_PACKAGE to '1'. Else,
RECOVRY_IMAGE to '1'(The GENERATE_AB_OTA_PACKAGE flag would not matter
in this case; enabling RECOVERY_IMAGE would mean a non A/B target and
hence will by default generate a non A/B OTA package).

Also, introduce a list of emmc/nand targets that support OTA upgrade
currently. OTA upgrade support on a target would be controlled by this
list and not the target-specific .conf file.

Change-Id: I7c154b5c1477376a3aaf7b577d76763621937698