edk2: remove weird hack to sstate_task_prefunc()
For unknown reasons, this recipe had been overroding the default sstate cache
management hook. The overridden copy caused untracked files to pollute
DEPLOY_DIR, which leads to conflicts any time that the edk2 recipe got rebuilt
because Bitbake judged it to be out of date.

There was no value in the override. We have verified that with the hook
removed, exactly the same payload is put into DEPLOY_DIR on an initial
execution of the recipe (without a hot sstate-cache) and from a copy of the
payload pulled from an up-to-date sstate-cache.

Remove it. Now rebuilds don't generate conflicts from attempting to execute
do_deploy() and place results into DEPLOY_DIR which conflict from untracked
files deposited there by the previous iteration's do_deploy().

Change-Id: I7a9e1c7951938114fe2e70b40e45c0cdc757e9f2